+++ lib/src/public_suffix_base.dart
@@ -12,11 +12,12 @@ import 'package:punycode/punycode.dart';
 
 import 'suffix_rules.dart';
 
-/// A description of the public suffix, root domain and registrable domain for a URI.
+/// A description of the public suffix, root domain and registrable domain for a URL.
 class PublicSuffix {
   final Uri sourceUri;
 
   bool _sourcePunycoded = false;
+  bool _hasKnownSuffix;
   String _root;
   String _suffix;
   String _domain;
@@ -28,52 +29,52 @@ class PublicSuffix {
 
   PublicSuffix _punyDecoded;
 
-  /// Returns the registrable domain part of the URI, based on both ICANN/IANA and private rules.
+  /// Returns the registrable domain part of the URL, based on both ICANN/IANA and private rules.
   ///
   /// The registrable domain is the public suffix and one preceding label.
   /// For example, `images.google.co.uk` has the registrable domain `google.co.uk`.
   String get domain => _domain;
 
-  /// Returns the subdomain part of the URI, based on both ICANN/IANA and private rules.
+  /// Returns the subdomain part of the URL, based on both ICANN/IANA and private rules.
   ///
   /// The subdomain is the part of the host that precedes the registrable domain
   /// (see [domain]).
   /// For example, `images.google.co.uk` has the subdomain `images`.
   String get subdomain => _subdomain;
 
-  /// Returns the root domain part of the URI, based on both ICANN/IANA and private rules.
+  /// Returns the root domain part of the URL, based on both ICANN/IANA and private rules.
   ///
   /// The root domain is the label that precedes the public suffix.
   /// For example, `images.google.co.uk` has the root domain `google`.
   String get root => _root;
 
-  /// Returns the public suffix part of the URI, based on both ICANN/IANA and private rules.
+  /// Returns the public suffix part of the URL, based on both ICANN/IANA and private rules.
   ///
   /// The public suffix is the labels at the end of the URL which are not controlled
   /// by the registrant of the domain.
   /// For example, `images.google.co.uk` has the public suffix `co.uk`.
   String get suffix => _suffix;
 
-  /// Returns the registrable domain part of the URI, based on ICANN/IANA rules.
+  /// Returns the registrable domain part of the URL, based on ICANN/IANA rules.
   ///
   /// The registrable domain is the public suffix and one preceding label.
   /// For example, `images.google.co.uk` has the registrable domain `google.co.uk`.
   String get icannDomain => _icannDomain;
 
-  /// Returns the subdomain part of the URI, based on ICANN/IANA rules.
+  /// Returns the subdomain part of the URL, based on ICANN/IANA rules.
   ///
   /// The subdomain is the part of the host that precedes the registrable domain
   /// (see [domain]).
   /// For example, `images.google.co.uk` has the subdomain `images`.
   String get icannSubdomain => _icannSubdomain;
 
-  /// Returns the root domain part of the URI, based on ICANN/IANA rules.
+  /// Returns the root domain part of the URL, based on ICANN/IANA rules.
   ///
   /// The root domain is the label that precedes the public suffix.
   /// For example, `images.google.co.uk` has the root domain `google`.
   String get icannRoot => _icannRoot;
 
-  /// Returns the public suffix part of the URI, based on ICANN/IANA rules.
+  /// Returns the public suffix part of the URL, based on ICANN/IANA rules.
   ///
   /// The public suffix is the labels at the end of the URL which are not controlled
   /// by the registrant of the domain.
@@ -83,11 +84,55 @@ class PublicSuffix {
   /// Returns a punycode decoded version of this object.
   PublicSuffix get punyDecoded => _punyDecoded;
 
-  /// Checks if the URI was matched with a private rule rather than an ICANN/IANA rule.
+  /// Checks if the URL was matched with a private rule rather than an ICANN/IANA rule.
   ///
   /// If [true], then [root], [suffix] and [domain] will be different from the
   /// `icann`-prefixed getters.
   bool isPrivateSuffix() => icannSuffix != _suffix;
+
+  /// Whether the [suffix] is a known suffix or not.
+  ///
+  /// A known suffix is one which has a rule in the suffix rule list.
+  bool hasKnownSuffix() => _hasKnownSuffix;
+
+  /// Checks if the registrable domain is valid.
+  ///
+  /// If [icann] is [true] the check will be based on [icannDomain], otherwise
+  /// [domain] is used.
+  /// If [acceptDefaultRule] is [false] URLs with suffixes only matching the
+  /// default rule (`*`) will be seen as invalid.
+  bool hasValidDomain({bool icann = false, bool acceptDefaultRule = true}) {
+    var _domain = (icann ? icannDomain : domain);
+
+    if (acceptDefaultRule || hasKnownSuffix()) {
+      return _domain != null;
+    } else {
+      return false;
+    }
+  }
+
+  /// Checks if this object represents a subdomain of another.
+  ///
+  /// The domain and subdomain properties are compared to determine if
+  /// this object represents a subdomain of [other]. If [icann] is [true],
+  /// comparison will be based on only the ICANN/IANA rules.
+  ///
+  /// For example, `http://images.google.co.uk` is a subdomain of `http://google.co.uk`.
+  ///
+  /// If [other] has a subdomain and this object represents a subdomain of that,
+  /// [true] is still returned.
+  bool isSubdomainOf(PublicSuffix other, {bool icann = false}) {
+    if (icann) {
+      return icannDomain == other.icannDomain &&
+          icannSubdomain != null &&
+          (other.icannSubdomain == null ||
+              icannSubdomain.endsWith(other.icannSubdomain));
+    } else {
+      return domain == other.domain &&
+          subdomain != null &&
+          (other.subdomain == null || subdomain.endsWith(other.subdomain));
+    }
+  }
 
   PublicSuffix._(this.sourceUri, String host, this._root, this._suffix,
       this._icannRoot, this._icannSuffix) {
@@ -109,14 +154,20 @@ class PublicSuffix {
     }
     if (!sourceUri.hasAuthority) {
       throw ArgumentError(
-          "The URI is missing the authority component: $sourceUri");
+          "The URL is missing the authority component: $sourceUri");
     }
 
-    _parseUri(sourceUri, SuffixRules.ruleMap);
+    _parseUrl(sourceUri, SuffixRules.ruleMap);
   }
 
-  void _parseUri(Uri uri, Map<String, Iterable<Rule>> suffixMap) {
-    var host = _decodeHost(uri);
+  /// Creates a new instance from a URL in a string.
+  ///
+  /// This is a convenience method that simply converts [url] into a URI object
+  /// and creates an instance from it.
+  PublicSuffix.fromString(String url) : this(Uri.parse(url));
+
+  void _parseUrl(Uri url, Map<String, Iterable<Rule>> suffixMap) {
+    var host = _decodeHost(url);
     var matchingRules = _findMatchingRules(host, suffixMap);
     var prevailingIcannRule = _getPrevailingRule(matchingRules['icann']);
     var prevailingAllRule = _getPrevailingRule(matchingRules['all']);
@@ -140,6 +191,7 @@ class PublicSuffix {
     _icannRoot = icannData['root'];
     _icannDomain = icannData['registrable'];
     _icannSubdomain = icannData['sub'];
+    _hasKnownSuffix = (prevailingAllRule.labels != "*");
 
     var puny = allData['puny'];
     var icannPuny = icannData['puny'];
@@ -148,8 +200,8 @@ class PublicSuffix {
         icannPuny['root'], icannPuny['suffix']);
   }
 
-  String _decodeHost(Uri uri) {
-    var host = uri.host.replaceAll(RegExp(r'\.+$'), '').toLowerCase();
+  String _decodeHost(Uri url) {
+    var host = url.host.replaceAll(RegExp(r'\.+$'), '').toLowerCase();
     host = Uri.decodeComponent(host);
 
     var punycodes = RegExp(r'xn--[a-z0-9-]+').allMatches(host);
@@ -178,7 +230,7 @@ class PublicSuffix {
 
     if (suffixList != null) {
       for (var rule in suffixList) {
-        if (_ruleMatches(rule, host)) {
+        if (rule.matches(host)) {
           allMatches.add(rule);
 
           if (rule.isIcann) {
@@ -189,37 +241,6 @@ class PublicSuffix {
     }
 
     return {'icann': icannMatches, 'all': allMatches};
-  }
-
-  bool _ruleMatches(Rule rule, String host) {
-    var hostParts = host.split('.');
-    var ruleParts = rule.labels.split('.');
-
-    hostParts.removeWhere((e) => e.isEmpty);
-
-    var matches = true;
-
-    if (ruleParts.length <= hostParts.length) {
-      int r = ruleParts.length - 1;
-      int h = hostParts.length - 1;
-
-      while (r >= 0) {
-        var rulePart = ruleParts[r];
-        var hostPart = hostParts[h];
-
-        if (rulePart != '*' && rulePart != hostPart) {
-          matches = false;
-          break;
-        }
-
-        r--;
-        h--;
-      }
-    } else {
-      matches = false;
-    }
-
-    return matches;
   }
 
   Rule _getPrevailingRule(List<Rule> matchingRules) {